requirements.txt
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==44.0.0
et_xmlfile==2.0.0
fpdf==1.7.2
idna==3.10
lxml==5.3.0
msal==1.31.1
numpy==2.1.3
Office365-REST-Python-Client==2.5.14
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
pyodbc==5.2.0
PySide6==6.8.1
PySide6_Addons==6.8.1
PySide6_Essentials==6.8.1
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
qrcode==8.0
requests==2.32.3
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
schedule==1.2.2
SharePlum==0.5.1
shiboken6==6.8.1
six==1.16.0
SQLAlchemy==2.0.36
sspilib==0.2.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
